Center form ad the submit button
