+++ package.json
@@ -17,9 +17,11 @@
     "react-datepicker": "^4.5.0",
     "react-dom": "^17.0.2",
     "react-hook-form": "^7.22.2",
+    "react-icons": "^4.3.1",
     "react-redux": "^7.2.6",
     "react-router-dom": "^6.1.1",
     "react-scripts": "^5.0.0",
+    "react-select": "^5.2.1",
     "redux": "^4.1.2",
     "redux-devtools-extension": "^2.13.9",
     "redux-thunk": "^2.4.1",
