--- src/components/Reserve.js
@@ -1,5 +1,0 @@
-import React from 'react';
-
-const Reserve = () => (<h1>Reserve</h1>);
-
-export default Reserve;
