+++ src/sass/main.scss
@@ -1,8 +1,2 @@
 
 @import './reservation'
-
-* {
-    margin: 0;
-    padding: 0;
-    box-sizing: border-box;
-}
